requirements.txt
Flask
dnspython
